requirements.txt
Flask
dnspython
